add caching for LlrState
